requirements.txt
Flask==3.0.3
ffmpeg-python==0.2.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
yt-dlp==2023.12.30
yt-dlp==2023.12.30
